added catch to DELETE ensuring the image path in SQl matches path we have selected
https://app.asana.com/0/161555263951748/169698496837386

+++ public_html/php/apis/image/index.php
@@ -162,10 +162,13 @@ try {
 				throw(new RuntimeException("The image does not exist!"));
 			}
 
-			$image->getImageFileName(); //wtf do i need this for??
+			if(($requestObject->fileName) === ($image->getImageFileName())){
 
-			$image->delete($pdo);
-			$reply->message("Image deleted A-OK");
+				$image->delete($pdo);
+				$reply->message("Image deleted A-OK");
+
+			}
+
 
 		} else {
 			throw (new InvalidArgumentException("Invalid HTTP method request"));
